Move RetryManager to foam.util.retry package
